LayoutTests/http/tests/websocket/tests/hybi/secure-cookie-insecure-connection-expected.txt

Test WebSocket does not send Secure cookies over an insecure connection.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

WebSocket open
WebSocket closed
PASS cookie is "WK-websocket-test=1"
PASS successfullyParsed is true

TEST COMPLETE
